Cosmetic change to philox and threefry
Arguments are [c0, c1, c2, c3, k0, k1]
rather than   [k0, k1, c0, c1, c2, c3]

This means the "counter" in the counter_based_engine map to one of the
"counter" arguments in the original paper, which seems helpful for
exposition, but is purely cosmetic.

It also makes it slightly easier to use the original paper's
kat_vectors.